Forgot one thing

+++ gl_testbench/Dx12Mesh.h
@@ -36,7 +36,7 @@ public:
 	void setTranslation(float3 translation);
 
 	void createMeshFromObj(const wchar_t* filename);
-	void createMesh(float* meshPos, float* meshNor, float* meshUV, unsigned long* meshInd, size_t numVert, size_t numInd);
+	void createMesh(float* meshPos, float* meshNor, float* meshUV, DWORD* meshInd, size_t numVert, size_t numInd);
 	void createCube();
 	void createTriangle();
 	void createQuad();
